% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_size.R
\name{object_size}
\alias{object_size}
\alias{as.object_size}
\title{Utility function to measure object size}
\usage{
object_size(x, unit = "MB", standard = "SI")

as.object_size(x, unit = "MB", standard = "SI")
}
\arguments{
\item{x}{an \R object.}

\item{standard}{the byte-size unit standard to be used.  A character
    string, possibly abbreviated from \code{"legacy"}, \code{"IEC"},
    \code{"SI"} and \code{"auto"}.  See \sQuote{Formatting and printing
      object sizes} for details.}
}
\value{
A scalar character vector of the input object's size in the units
requested
}
\description{
\code{object_size} works on an object and provides the size of the
object using \code{utils::object.size}. \code{as.object_size} converts a numeric
value to the units indicated in the \code{unit} argument.
}
\examples{
var <- 1e10
object_size(var, unit = "KB")

as.object_size(2500000)
}
